Move and update the cfg reading code to Rust
The Rust config code right now is called by cfg.cpp which then
loads the INI and updates the cfg global in CPP. This is clunky
but a nice cleanup is coming when we move the code that relies
on cfg.cpp to Rust. At that point that code should also be cleaner.

This code includes some tests with random configurations (need to
add more), and a lot of refactor around the loading of different
sequences of configs (like aspect ratios and player controller).

As far as I know, this is backward compatible, but I need more
tests. It is also very difficult to test against the CPP code
as I need to run it on the DE10-Nano itself. I don't think
I really need to add harnesses or try to build the CPP code
on desktop, all I need is to add a new executable target with
a lot of different examples INI to try to find differences.