Entirely fix the INI config parsing and loading
Now loading from Rust is 100% byte compatible with the old config.
There are just video overrides that wont load, but working on it.

This is the most insane part of the code (least straightforward).
The INI loading is custom made and not entirely standard INI. It
also has a lot of side effects and edge cases which made using a
out-of-the-box INI loading library impossible. So I had to move
the INI to JSON then load it using a JSON loader.

At least it works now, so moving to supporting overrides and
then to fpga specific code.