Add two new projects (senior and junior) to help manage MiSTer remotely
They are both based on the same technology as GoLEm, but senior is a
REST API server that can install cores and run low level commands,
and junior is a CLI that can send commands to senior.

Junior is currently not started. I only need senior for my debugging
needs.