Implement the new JavaScript frontend for Golem (#30)
# New UI Code

The whole UI is not using JavaScript for its flow and logic. That is, the menus and panels are still implemented in Rust (for performance), but the content of those menus are in JavaScript. This allows us to quickly iterate on menus, something that was kind of difficult before.

The performance is unnoticeable, after all the DE10 is a 800Mhz dual-core processor, more than enough to run a full web browser in memory.

# Performance in Core

There is no change when the core is running. The JavaScript is only responsible to give the path and information to GoLEm for running the core, then GoLEm takes over and go into its core loop. The Core menu is currently unchanged, but we could see using JavaScript in there too. We'll see, for now the core menu will stay the same.

# Thank you to Patreons

I added an About menu that shows a thank you to all current paying patrons, and contributors. This list can be changed if you want to update your name.