Many changes under the hood, mostly QoL (#39)
This started as an experiment to use the framebuffer from the MENU core to display a full RGB menu. It worked, but was not performant enough to justify going down that route.

During the creation of this PR a lot of improvements to the internals made the code much more maintainable and will help with future development.